further refactor tests for performance
